move remove avatar modal to separate file
