+++ packages/client/src/pages/admin2/components/avatar/RemoveAvatarModal.tsx
@@ -1,0 +1,58 @@
+/*
+CPAL-1.0 License
+
+The contents of this file are subject to the Common Public Attribution License
+Version 1.0. (the "License"); you may not use this file except in compliance
+with the License. You may obtain a copy of the License at
+https://github.com/EtherealEngine/etherealengine/blob/dev/LICENSE.
+The License is based on the Mozilla Public License Version 1.1, but Sections 14
+and 15 have been added to cover use of software over a computer network and 
+provide for limited attribution for the Original Developer. In addition, 
+Exhibit A has been modified to be consistent with Exhibit B.
+
+Software distributed under the License is distributed on an "AS IS" basis,
+WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License for the
+specific language governing rights and limitations under the License.
+
+The Original Code is Ethereal Engine.
+
+The Original Developer is the Initial Developer. The Initial Developer of the
+Original Code is the Ethereal Engine team.
+
+All portions of the code written by the Ethereal Engine team are Copyright © 2021-2023 
+Ethereal Engine. All Rights Reserved.
+*/
+
+import { PopoverState } from '@etherealengine/client-core/src/common/services/PopoverState'
+import { AvatarType, avatarPath } from '@etherealengine/common/src/schema.type.module'
+import { useMutation } from '@etherealengine/spatial/src/common/functions/FeathersHooks'
+import LoadingCircle from '@etherealengine/ui/src/primitives/tailwind/LoadingCircle'
+import Modal from '@etherealengine/ui/src/primitives/tailwind/Modal'
+import Text from '@etherealengine/ui/src/primitives/tailwind/Text'
+import { useHookstate } from '@hookstate/core'
+import React from 'react'
+import { useTranslation } from 'react-i18next'
+
+export default function RemoveAvatarModal({ avatar }: { avatar: AvatarType }) {
+  const { t } = useTranslation()
+  const adminAvatarRemove = useMutation(avatarPath).remove
+  const modalProcessing = useHookstate(false)
+
+  return (
+    <Modal
+      title={t('admin:components.avatar.remove')}
+      onSubmit={() => {
+        modalProcessing.set(true)
+        adminAvatarRemove(avatar.id)
+      }}
+      onClose={!modalProcessing.value ? () => PopoverState.hidePopupover() : undefined}
+      hideFooter={modalProcessing.value}
+    >
+      {modalProcessing.value ? (
+        <LoadingCircle className="h-[10vh]" />
+      ) : (
+        <Text>{`${t('admin:components.avatar.confirmAvatarDelete')} '${avatar.name}'?`}</Text>
+      )}
+    </Modal>
+  )
+}
